Move BufferWriter back to searchlib.
