+++ searchcore/src/vespa/searchcore/proton/documentmetastore/documentmetastoresaver.cpp
@@ -1,7 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "documentmetastoresaver.h"
-#include <vespa/vespalib/util/bufferwriter.h>
+#include <vespa/searchlib/util/bufferwriter.h>
 #include "document_meta_store_versions.h"
 #include <vespa/searchlib/attribute/iattributesavetarget.h>
