+++ searchlib/src/vespa/searchlib/attribute/singleenumattributesaver.cpp
@@ -2,7 +2,7 @@
 
 #include "singleenumattributesaver.h"
 #include <vespa/vespalib/util/array.hpp>
-#include <vespa/vespalib/util/bufferwriter.h>
+#include <vespa/searchlib/util/bufferwriter.h>
 #include "iattributesavetarget.h"
