+++ searchlib/src/vespa/searchlib/attribute/singlestringattribute.hpp
@@ -9,7 +9,7 @@
 #include <vespa/fastlib/io/bufferedfile.h>
 #include <vespa/vespalib/text/utf8.h>
 #include <vespa/vespalib/text/lowercase.h>
-#include <vespa/vespalib/util/bufferwriter.h>
+#include <vespa/searchlib/util/bufferwriter.h>
 #include <vespa/vespalib/util/regexp.h>
 #include <vespa/searchlib/query/query_term_ucs4.h>
