+++ searchlib/src/vespa/searchlib/util/CMakeLists.txt
@@ -1,6 +1,7 @@
 # Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchlib_util OBJECT
     SOURCES
+    bufferwriter.cpp
     comprbuffer.cpp
     comprfile.cpp
     dirtraverse.cpp
