+++ searchlib/src/vespa/searchlib/util/bufferwriter.cpp
