+++ searchlib/src/vespa/searchlib/util/bufferwriter.h
