+++ vespalib/src/vespa/vespalib/datastore/unique_store.hpp
@@ -10,7 +10,6 @@
 #include "unique_store_builder.hpp"
 #include "unique_store_dictionary.hpp"
 #include "unique_store_enumerator.hpp"
-#include <vespa/vespalib/util/bufferwriter.h>
 #include <atomic>
 #include <algorithm>
